rename: Strategy를 Strategy package로 분리
