+++ src/test/java/racingcar/strategy/RandomStrategyTest.java
@@ -1,0 +1,20 @@
+package racingcar.strategy;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class RandomStrategyTest {
+
+  @Test
+  void isMovableTrue() {
+    MoveStrategy moveStrategy = () -> true;
+    assertThat(moveStrategy.isMovable()).isTrue();
+  }
+
+  @Test
+  void isMovableFalse() {
+    MoveStrategy moveStrategy = () -> false;
+    assertThat(moveStrategy.isMovable()).isFalse();
+  }
+}
